exponentially make requests done :)

+++ project/parser.h
@@ -17,7 +17,7 @@
 typedef struct ParsedFile {
     int busCapacity;
     int boardingTime;
-    float requestRate;
+    double requestRate;
     float pickupInterval;
     int maxDelay;
     int noBuses;
